Add files via upload
code to generate wavelet features and already generated features themselves (takes a while to run). The features generated are organized by RASS number and which sensor ("axis [1-3]"). Vector magnitudes were used as opposed to raw signals.

Feature generation used the cwt() continuous wavelet function with a morse mother wavelet (yielding a 2D complex-valued matrix where each row of the output represents correlation of original signal with a different scaling factor applied to the wavelet), the cwt's were split up by RASS score and for each section of code findpeaks was used to find the magnitude of correlation where there was a correlation peak, the scaling factor at that peak, the width of the peak, and the prominence of the peak respectively.

In simpler terms, the first two columns show how well that section of RASS #N correlates with different scaling factors (by finding areas of peak correlation) and the latter two give some idea of the morphology of those areas of peak correlation